Move the 7zip helper functions out of ZipEncryptionHandling into their own class, for easier use elsewhere
